Set up the framework of the landing page's content
This version of the project sets up the inital
structure of the landing page using div elements.
I expect that the structure will be modified as
the project evolves.

The CSS styling for the landing page is yet to be
implemented. That will come in future iterations
of the project.