Arrange the content in the call-to-action section
I used certain flexbox properties and
declarations to arrange the content in the
"call-to-action" section of the webpage based on
the landing page example provided in the Odin
Project's project page.

Also, I modified the HTML file to include more
div containers for my flexbox layout.

Lastly, I fixed some typos or renamed certain
elements to better suit the naming convention
used in the code.

Please note, I am using borders on all the div
and ul elements to visually see how the flexbox
properties and declarations are affecting my flex
items when using flexbox. I will be removing
these borders when I am done with arranging items
with flexbox and ready to style the webpage with
general CSS properties.