Apply CSS styling to the webpage based on requirements
This version of the landing page code contains
some box model and general CSS styling to "match"
the styling in the image file titled
"landing-page-example.png".

Also, I applied some more div containers in the
header of the landing page to help with styling
the logo.

Please note, I am using borders on all the div
and ul elements to visually see how the flexbox
properties and declarations are affecting my flex
items when using flexbox. I will be keeping them
until I am fairly confident that the styling like
font-size and font-weight does not negatively
impact the layout of the content in the landing
page.